requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pymongo==4.6.1
motor==3.3.2
pydantic==2.5.3
pydantic-settings>=2.0.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.2.0
cryptography==42.0.0
streamlit==1.31.0
requests==2.31.0
aiosmtplib==3.0.1
email-validator==2.1.0
redis==5.0.1